venueChat and privateChats refactoring
