+++ src/components/organisms/ChatSidebar/components/VenueChat/VenueChat.tsx
@@ -5,7 +5,7 @@ import { AnyVenue } from "types/venues";
 
 import { WithId } from "utils/id";
 
-import { useVenueChat } from "hooks/useVenueChat";
+import { useVenueChat } from "hooks/chats/venueChat";
 
 import { Chatbox } from "components/molecules/Chatbox";
 
@@ -19,7 +19,7 @@ export const _VenueChat: React.FC<VenueChatProps> = ({ venue }) => {
   const {
     sendMessage,
     deleteMessage,
-    messagesToDisplay,
+    messages,
     sendThreadReply,
   } = useVenueChat(venue.id);
 
@@ -28,7 +28,7 @@ export const _VenueChat: React.FC<VenueChatProps> = ({ venue }) => {
       <Chatbox
         // poll is available for Venue Chat only (displayPoll = true)
         displayPoll
-        messages={messagesToDisplay}
+        messages={messages}
         sendMessage={sendMessage}
         sendThreadReply={sendThreadReply}
         deleteMessage={deleteMessage}
