+++ src/components/templates/ReactionPage/ReactionPage.tsx
@@ -8,7 +8,7 @@ import { useConnectCurrentVenueNG } from "hooks/useConnectCurrentVenueNG";
 import { useFirestoreConnect } from "hooks/useFirestoreConnect";
 import { useSelector } from "hooks/useSelector";
 import { useRecentVenueUsers } from "hooks/users";
-import { useVenueChat } from "hooks/useVenueChat";
+import { useVenueChat } from "hooks/chats/venueChat";
 import { useVenueId } from "hooks/useVenueId";
 
 import { UserList } from "components/molecules/UserList";
@@ -28,7 +28,7 @@ export const ReactionPage: React.FC = () => {
   const { recentVenueUsers } = useRecentVenueUsers({
     venueName: currentVenue?.name,
   });
-  const { messagesToDisplay: venueChatMessages } = useVenueChat(venueId);
+  const { messages: venueChatMessages } = useVenueChat(venueId);
 
   // @debt this is very similar to the query in src/hooks/reactions.tsx, but that filters by createdAt > now
   useFirestoreConnect(
